CRA-9 Fixed bug in memory estimation. Memory size of one grid was measured as int (due to implicit casting) rather than long int.
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@638 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/model.cpp
@@ -494,7 +494,7 @@ Model::checkAvailableMemory(Simbox        * timeSimbox,
                                     modelSettings->getNXpad(), 
                                     modelSettings->getNYpad(), 
                                     modelSettings->getNZpad());
-  long long int gridSizePad = 4*dummyGrid->getrsize();
+  long long int gridSizePad = static_cast<long long int>(4)*dummyGrid->getrsize();
 
   delete dummyGrid;
   dummyGrid = new FFTGrid(timeSimbox->getnx(),
